[AIRFLOW-3469] Move KnownEvent out of models.py
